feat(train_yolo_transcaranet.sh): add script to train YOLO TransCaranet model and notify via Slack if training fails
The commit adds a new file `train_yolo_transcaranet.sh` which contains a script to train the YOLO TransCaranet model. The script is executed using Python 3 and calls the `train_trans_caranet.py` script. In case the training fails, the script also includes a fallback option to execute the `slack_bot.py` script from the `tools` directory to notify via Slack. This addition allows for easier training of the YOLO TransCaranet model and provides a notification mechanism in case of failures.