chore(polyp491.yaml): update dataset paths and remove unnecessary comments
The dataset paths in the polyp491.yaml file were updated to use a relative path and remove unnecessary comments. The 'path' variable was changed to './dataset_v0' to specify the dataset root directory. The 'train' and 'val' variables were updated to remove the '/images' suffix as it is already specified in the 'path' variable. The 'test' variable was left unchanged.