chore(polyp491.yaml): update dataset paths to reflect new directory structure
The dataset paths in the polyp491.yaml file have been updated to reflect the new directory structure. The 'path' variable has been changed to '../datasets/dataset_v0/sekkai' and the 'train', 'val', and 'test' variables have been updated to 'images/sekkai_TrainDataset', 'images/sekkai_ValDataset', and 'images/sekkai_TestDataset' respectively. This change was made to ensure that the dataset paths are accurate and up to date.